requirements.txt
datefinder
pytesseract
regex
argparse
pillow
